refactor : home login에서 error with buttons 이름 변경 및 기능 추가 적용
