+++ src/components/ErrorWithButtons/ErrorWithButtons.style.tsx
