+++ src/pages/NotFound/index.tsx
@@ -1,8 +1,8 @@
-import HomeLogin from '@components/HomeLogin';
+import ErrorWithButtons from '@components/ErrorWithButtons';
 import { wrongApproachMEntion } from '@constants/mentions';
 
 const NotFound = () => {
-  return <HomeLogin mention={wrongApproachMEntion} />;
+  return <ErrorWithButtons mention={wrongApproachMEntion} />;
 };
 
 export default NotFound;
