+++ src/router/ProtectedRoute.tsx
@@ -1,7 +1,7 @@
 import { Outlet } from 'react-router-dom';
 import { useRecoilValueLoadable, useSetRecoilState } from 'recoil';
 
-import HomeLogin from '@components/HomeLogin';
+import ErrorWithButtons from '@components/ErrorWithButtons';
 import Loading from '@components/Loading';
 import { portalState } from '@store/portal';
 import { isLoginState } from '@store/user';
@@ -14,10 +14,10 @@ const ProtectedRoute = () => {
     switch (state) {
       case 'hasValue':
         if (!contents) setPortal('login');
-        return contents ? <Outlet /> : <HomeLogin />;
+        return contents ? <Outlet /> : <ErrorWithButtons />;
       case 'hasError':
         setPortal('login');
-        return <HomeLogin />;
+        return <ErrorWithButtons />;
       case 'loading':
         return <Loading color='grey1' size={42} border={6} height='100vh' />;
     }
